requirements.txt
openpyxl
pandas
requests
selenium
webdriver_manager
python-dotenv
bs4